+++ Routeur/Classes/GridRouter/RoutingGridPoint.vb
@@ -412,7 +412,7 @@ Public Class RoutingGridPoint
         End Set
     End Property
 
-    Public Function GetHashCode1() As Integer Implements ICoords.GetHashCode
+    Public Function GetHashCode1() As Long Implements ICoords.GetHashCode
         Return _P.P.GetHashCode1
     End Function
 End Class
