+++ Routeur/Classes/ListMgr/ICoords.vb
@@ -4,6 +4,6 @@
     Property Lat() As Double
 
     Function Equals(ByVal C As ICoords) As Boolean
-    Function GetHashCode() As Integer
+    Function GetHashCode() As Long
 
 End Interface
